+++ .openshiftio/application.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Template
 metadata:
-  name: spring-boot-health-check
+  name: health-check
   annotations:
     iconClass: icon-spring
     tags: spring-boot, health check, java, microservice
@@ -50,7 +50,7 @@ objects:
 - apiVersion: v1
   kind: ImageStream
   metadata:
-    name: spring-boot-health-check
+    name: health-check
   spec: {}
 - apiVersion: v1
   kind: ImageStream
@@ -65,12 +65,12 @@ objects:
 - apiVersion: v1
   kind: BuildConfig
   metadata:
-    name: spring-boot-health-check
+    name: health-check
   spec:
     output:
       to:
         kind: ImageStreamTag
-        name: spring-boot-health-check:BOOSTER_VERSION
+        name: health-check:BOOSTER_VERSION
     postCommit: {}
     resources: {}
     source:
@@ -107,11 +107,11 @@ objects:
   kind: Service
   metadata:
     labels:
-      app: spring-boot-health-check
+      app: health-check
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-health-check
+      group: dev.snowdrop.example
+    name: health-check
   spec:
     ports:
     - name: http
@@ -119,25 +119,25 @@ objects:
       protocol: TCP
       targetPort: 8080
     selector:
-      app: spring-boot-health-check
+      app: health-check
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
     labels:
-      app: spring-boot-health-check
+      app: health-check
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-health-check
+      group: dev.snowdrop.example
+    name: health-check
   spec:
     replicas: 1
     revisionHistoryLimit: 2
     selector:
-      app: spring-boot-health-check
+      app: health-check
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -145,10 +145,10 @@ objects:
     template:
       metadata:
         labels:
-          app: spring-boot-health-check
+          app: health-check
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -156,7 +156,7 @@ objects:
             valueFrom:
               fieldRef:
                 fieldPath: metadata.namespace
-          image: spring-boot-health-check:BOOSTER_VERSION
+          image: health-check:BOOSTER_VERSION
           imagePullPolicy: IfNotPresent
           livenessProbe:
             failureThreshold: 2
@@ -197,20 +197,20 @@ objects:
         - spring-boot
         from:
           kind: ImageStreamTag
-          name: spring-boot-health-check:BOOSTER_VERSION
+          name: health-check:BOOSTER_VERSION
       type: ImageChange
 - apiVersion: v1
   kind: Route
   metadata:
     labels:
-      app: spring-boot-health-check
+      app: health-check
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-health-check
+      group: dev.snowdrop.example
+    name: health-check
   spec:
     port:
       targetPort: 8080
     to:
       kind: Service
-      name: spring-boot-health-check
+      name: health-check
