+++ README.md
@@ -1,3 +1,3 @@
-[![CircleCI](https://circleci.com/gh/snowdrop/spring-boot-health-check-booster.svg?style=shield)](https://circleci.com/gh/snowdrop/spring-boot-health-check-booster)
+[![CircleCI](https://circleci.com/gh/snowdrop/health-check-example.svg?style=shield)](https://circleci.com/gh/snowdrop/health-check-example)
 
 https://appdev.openshift.io/docs/spring-boot-runtime.html#mission-health-check-spring-boot
